add expansion , but just simple paramtre expansion is work for the moment

+++ libft/additional_function/ft_max.c
@@ -6,16 +6,16 @@
 /*   By: aait-ihi <aait-ihi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/25 19:57:40 by aait-ihi          #+#    #+#             */
-/*   Updated: 2019/12/24 01:29:05 by aait-ihi         ###   ########.fr       */
+/*   Updated: 2020/02/19 18:55:32 by aait-ihi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_max(int a, int b)
+long	ft_max(long a, long b)
 {
 	return (a > b ? a : b);
 }
 
-int	ft_min(int a, int b)
+long	ft_min(long a, long b)
 {
 	return (a < b ? a : b);
 }
